deployment_info::contracts

Module ISafe

Module containing a contract’s types and functions.

interface ISafe {
    function VERSION() external view returns (string memory);
    function getOwners() external view returns (address[] memory);
    function getThreshold() external view returns (uint256);
}

Structs

ISafeInstance

A ISafe instance.

VERSIONCall

Function with signature VERSION() and selector 0xffa1ad74.

VERSIONReturn

Container type for the return parameters of the VERSION() function.

getOwnersCall

Function with signature getOwners() and selector 0xa0e67e2b.

getOwnersReturn

Container type for the return parameters of the getOwners() function.

getThresholdCall

Function with signature getThreshold() and selector 0xe75235b8.

getThresholdReturn

Container type for the return parameters of the getThreshold() function.

Enums

ISafeCalls

Container for all the ISafe function calls.

Functions

new

Creates a new wrapper around an on-chain ISafe contract instance.